backtest/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pandas==1.3.3
numpy==1.21.2
mlflow==2.13.0
python-dotenv==1.0.1
backtrader==1.9.78.123
redis==5.0.5
ta==0.11.0
